requirements.txt
gradio==3.50.2
numpy==1.26.4
Pillow==10.2.0
torch==2.2.0
torchvision==0.17.0
diffusers==0.19.3
transformers==4.30.0
accelerate==0.26.1
safetensors==0.4.2
pydantic==2.6.1
huggingface-hub==0.17.3
python-dotenv==1.0.0
requests==2.28.2
tiktoken==0.4.0
